services/auth-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.9
email-validator>=2.0.0